requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
python-dotenv==1.0.0
pydantic==2.5.0
supabase==2.3.0
transformers==4.36.0
